Copy table structure as is, don't add indexes manually
This removes the need to add indexes manually. It now
creates table with `INCLUDING ALL`, which copies
over the constraints, primary keys, indexes, storage and more. There
is a fix for serial/sequences on primary keys, to ensure
that both shadow table and primary don't use the same sequence.

Using `INCLUDING ALL` is a little better since it does
most of the work of transferring constraints, storage,
identity, etc for us. Which means, less state keeping
and having to support PG features manually also.

In case we see that having the indexes pre-created slows
down in performance when copying rows, we can look into alternate
more elaborate solutions. For now, this is a good starting point.

There is also a caveat that indexes, sequences and constraints
lose their original naming. Which I think is OK for now. Can be
supported in a new feature in future.