Support to perform operations in any schema
We were taking in schema as the input
but majority of the code written just
assumed public. This now sets the
search_path to the schema before performing
any operations.

Tweaks to specs and spec helpers to respect
the new settings and setting the search_path
accordingly in testing environment too