Validate constraints post swap
This way we validate the FKs that were added
w/ NOT VALID during the swap. This is a safe operation
to run post swap (since it doesn't block on the lock).
Also leaves the pg_get_constraintdef in a good state.
In other words, post this step, the FK's won't have NOT VALID,
that were added during the swap