Replaying remaining <20 rows (below delta) before swap
This ensures that when the delta count is < 20 rows (which indicates its time for swap),
it will replay those rows *after* the access exclusive lock
is acquired and then proceed to performing the swap.

This finally ensures no data loss - since we have already
acquired an access exclusive lock and can replay the <20 replays
swiftly and continue on with the process.

Changed Query.run api to support the use of existing
transaction such that lock is held