Acquire lock and attempt to kill backends when setting up trigger
Since we have the ability to kill competing backends (opt-in via a flag),
and additionally setting up trigger requires lock - this change attempts
to first acquire a lock and if kill-backends is supplied then it will kill the
backends and retry again to setup the triggers. If no-kill-backends is provided
it will try upto 3 times and fail.

This also means we will attempt to try and acquire access exclusive lock twice -
once when setting up triggers and once during swap. Accordingly kill backends (if supplied).

Added specs and tested