Pass custom copy statement for copy_data (#33)
This allows users to pass a custom sql file that will be played
when copying data from primary to shadow table.

Users will have full control over the operation and pg-osc
will just play the contents of the file. It takes in an optional
--copy-statement flag with the location of the sql file. There
is a template var that points to shadow_table which is replaced
when the query is being perform. This way users can customize
the 'INSER INTO' statement as well.

This is an advanced feature. There are some good uses like backfilling
that can be performed with this. h/t to Jeff Frost for this idea.

Freshened the README and added some examples