Ensure table names are unique and preset
Since we are not restoring index/sequence name for now,
ensuring the table names are unique so that if we run the same
operation on the same table, there is no sequence conflict. The sequence
will get overriden on the table anyways, but won't fail and complain
when trying to create a shadow before the swap. This is temporary until
we support restoring original index and sequence names. Can be a pro feature too