Use ISOLATION LEVEL SERIALIZABLE (#42)
This ensures that after setting up the triggers,
all other operations happen in serializable
until copy contents. This is to make sure there
are no race conditions between setting up
the trigger and begining to copy data (INSERT INTO)
into the shadow table. If we don't do it via serializable
or holding an extended access exclusive lock, there is a possibility
of some rows getting added to the audit table and also get copied
with copy_data and then later get replayed, causing duplication.

This makes the code handling a bit unpretty and is hard to test for.
In future, a better way to address the re-use of transaction w/o calling
BEGIN redundantly would be nice.