Introduce configurable PULL_BATCH_COUNT and DELTA_COUNT
On tables with high volume workloads,
having the ability to tune PULL_BATCH_COUNT
and DELTA_COUNT is nice for faster catchup and swap.

The default values are 1000 and 20 respectively. Which means,
in one pass pg-osc will replay 1000 rows and if the count is <20
then it will perform the swap and replay.

By introducing new cli flags (--pull-batch-count, --delta-count),
these figures can be tuned further