Delete items by audit table PK when replaying
This has few changes

- Once it replays items from the audit table, it deletes
those rows after. However, it was deleting rows by the PK
of the primary table and not the audit table, which created
a race condition on high volume tables.
- Moves up setting alter statement refresh commands in the Store
to avoid any risk of leaky transactions in the end

Tested this manually using pgbench and no longer seeing
any mis match between data post pg-osc run.

Will be bringing that integration test to CI, so we can run them
on every commit and catch regressions early