Add foreign keys to parent during swap (#67)
Until now we were structuraly copying the primary
table into shadow table. While doing so, we were also
copying the FKs. This could lead to locks on other queries
running on the primary/dependent table.

This PR fixes it by not adding the FKs to the shadow table,
but adding the FK to the parent table itself from the same transaction,
during the swap with 'NOT VALID'. It then, after swap, validates all constraint
on the self table (primary) and referential tables (other tables with FK of primary
table). So that, no FK constraints are left in NOT VALID state.

Add smoke tests using pgbench (--foreign-keys)