Update primary key sequence on shadow table (#72)
This ensures that the primary key sequence is consistent
on the shadow table, same as the primary table. Because, sequences
are unique in PG, we had to reset/rebuild the sequence on the shadow
table after copy. Doing this, we'd lose the PK sequence of nextval
and result in index/pk issues when writes are continued on the new table
after swap. This fixes that by updating the PK sequence on the shadow
table, so that after the swap the same sequence and nextval is continued.