Create shadow and audit with auatovacuum default (#97)
Calling an `ALTER` with a transaction open during serializable while we copy data onto the audit
table could potentially lead to lock queues. This avoids doing so by removing the `disable_vaccum`
function and taking advantage of `WITH` directive when setting up the shadow and audit table in the
first place. Thus avoiding the need to perform an `ALTER` on audit table or shadow table and avoid locking.