Introduce the ability to show estimate progress of copy (#146)
* Introduce the ability to show estimate progress of copy

Introducing a new `log_progress` function to estimate progress during data operations.
This function compares the sizes of the parent table and the shadow table using `pg_table_size`.
As we can't count the rows directly due to the ongoing transaction, `pg_table_size` provides the closest estimate.
Prior to this, we perform a vacuum and analyze on the parent table to enhance the accuracy of the size measurement.

It's important to note that the `pg_table_size` of the shadow table may not exactly match the parent table, especially for larger tables. This discrepancy is expected,
since the logging is a best effort measure. The function will terminate once `copy_data` completes as it watches for an instance var - `@copy_finished`

Progress is logged at 60-second intervals.

* Update spec to not use a transaction when running & vacuum

* Drop ruby 2.7 and using ruby 3.3.0

* Ruby 3.0.0 req in gemspec

* Show logs from the run in smoke specs

* Create shadow table outside of transaction

* Break early and move around function definition: